Move HmdaActor to parent package
