+++ persistence/src/main/scala/hmda/persistence/HmdaActor.scala
@@ -1,4 +1,4 @@
-package hmda.persistence.processing
+package hmda.persistence
 
 import akka.actor.{ Actor, ActorLogging }
